Lappy Backup

+++ Competitions/24th Apr/C.cpp
@@ -11,8 +11,8 @@ int main()
 
     // Looping for Each Test Case
     for (int test_case = 0; test_case < test_cases; test_case++) {
-        long int side; cin >> side;
-        cout << (side * (side)) + 2 * side + 2 << '\n';
+        long long int side; cin >> side;
+        cout << side * side + 2 * side + 2 << '\n';
     }
  
     return 0;
